correct en xkcd 1–199 with script, also some xkcd-side updates
Alright, so this one was kinda complex to deal with.

Some transcripts and alt texts were missing a double space from
the JSON. I was using a mix of bash’s and fish’s printf to make
the files, and I don’t know how they ended up making that mistake,
but they did. I also accidentally left in some shell escapes in
one.

There were some comics that were rotated due to EXIF metadata, but
apparently those got fixed, so I removed the _orient directories
and made some _revision ones instead. I don’t remember if I was
including the current revision in _revision, but it appeared
like I was, so I kept it.

This was hard to do on my phone, so maybe consider leaving a star
and smash that follow button, and maybe leave a message in the
issues above.