create test citation.cff file
this kinda sucks. the file spec says that the filename *needs* to be capitalized
as CITATION.cff to be valid, but that’s inconsistent with what I already have.
maybe I should capitalize README and COPYING and stuff anyway. GNU specifies COPYING as uppercase, but it’s not really a formal spec. anyway, I’m just making
sure GitHub will recognize this in
lowercase (it will)